Do not get item twice
Part of the story #12443 : Table: Minimal folder display

Go to a subfolder. Refresh the page. You should not have two call to GET /docman_items/:id

Before that three actions were performed:
- one to load the current folder title
  GET /docman_items/:id
- one to load the current folder content
  GET /docman_items/:id/docman_items
- one to load the current folder breadcrumbs
  GET /docman_items/:id/parents + GET /docman_items/:id

In order to save resources and rain forest, now the load of breadcrumbs is also
responsible of setting the current folder title. It saves a call to
GET /docman_items/:id.

Further refactoring steps are coming after this one to propose a better naming
of actions and state structure.

Change-Id: Ifb54f5b47b9de49bceb9f327021cd6ad3044fc77