Rename current_folder_parents
Part of the story #12443 : Table: Minimal folder display

My concern with "parents" is that it is excluding the current folder whereas
the latter is part of the collection. I hope that with the new proposal, it
will be less ambiguous, since current item is part of its hierarchy.

Rename breadcrumbs action in order to be more high level (the hierarchy is used
both for the breadcrumb and the title).

Furthermore, no need to have two different loading flags for the same thing:
remove is_loading_folder_title.

No functional changes.

Change-Id: I7a1dfb1cc3b9c536c1e0e653ecbf477a94fea272