Display a search box in the new docman
This is a part of the story #10724 : search in document view

It allows the user to search a file in the docman.

NOTE : The advanced mode of the search will be implemented later.

How to test:
 - Go to the plugin document via the new url
 - Fill the search box
 - Press enter.

You will be redirected to the old docman with the result of your search.

Change-Id: Ib7b29899d4848ba6f14b992b00df586a0637107b