Objects already uploaded for a repository should not be uploaded again
The batch API should not give actions to a client requesting to upload
a file already existing for the repository.

To test you can upload a file and then try to request the upload again
to the batch API, e.g. with curl:

    $ curl -X POST -H 'Accept: application/vnd.git-lfs+json' \
         'https://tuleap.example.com/plugins/git/projectname/reponame.git/info/lfs/objects/batch' \
         --data \
         '{"operation": "upload", "transfers": [ "basic" ], "objects": [{"oid": "12345678", "size": 123}]}'

This is part of story #12322: have git-lfs batch and basic transfer API

Change-Id: I49f4390c968c3d476a990872ecfbd073c9ce5181