Deploy a nginx configuration adapted to large file upload
Git LFS upload requests are no more blocked by the limit set
for the max body size of a request elsewhere in Tuleap.

A dedicated global endpoint /uploads/ is used to handle to large file
uploads instead of something specific to Git LFS. This endpoint
can be used by controllers capable to handle the body size of a request
themselves. Having something generic will allow us in the future to dedicate
this type of trafic to a dedicated FPM pool more easily and can be reused
outside of the context of Git LFS for anything that should handle large
and long lived requests (caused by I/O wait not by performance issue).

This is part of story #12322: have git-lfs batch and basic transfer API

Change-Id: I5030e7d003531ba9f26d6c50d4e7338ac8c76aca