Fix warning releated to PHP 7 in the field file
Issue can be reproduced by accessing an artifact with a field file
but without a changeste value associated to the field. You can
achieve that by creating an artifact from the REST route without
specifiying anything for the field file.

This is part of request #12147: Fix PHP 7.2 warnings and errors

Change-Id: I61343812f48dfa902d7cbe17704ff42e19fea0dd