Return the value of the link of a document
This is a part of the story #12444 Table : action on click

Update the REST route representation so that the route return the value of the link url of a document.
This link will be used later in the frontend of the new Docman.

How to test :
  - Make sure to have at least one link in your Docman
  - Go to  the API Explorer and test the route /docman_items/:id or /docman_items/:id/docman_items

If the document is not a link, you should have "link_properties = null".
If the document is a link , you should have "link_properties = your_link"

Change-Id: I6621f5f5afc146e1f43d98fab0d7ff5cf597b73d