Document manager upload endpoint can be used cross-origin like the REST API
This endpoint will be used after a call to the Tuleap REST API so
it should as accessible as the REST API otherwise cross origin client
will be broken.

To keep the Tus server as interoperable as possible, PSR-15 middlewares
are used to set the necessary headers. Sadly, Tuleap still supports PHP 5.6
so look alike interfaces are introduced in the meantime. Once the PHP 5.6
support is dropped, these interfaces should be swapped out for the PSR
interfaces.

How to access the test upload endpoint is described in the introductory
commit 426ca739d33e52787ad1439f52a5c8d658d56c8d.

This is part of story #12612: upload document via REST route

Change-Id: I47fc0e04c90a81f6fba039200d6c80c3f666b8ce