Unlock the max body size for large file uploads
A first tentative has been done in d9dafef376791857116bdf3829bc15d58f5311f5
but was incorrect.

nginx will match the /uploads/ location and then due to the usage of
try_files we will lost the settings we have set for the original location.

This is part of story #12327: set limits on git-lfs file usage

Change-Id: If5d6fd7adb6f43d71fdf7001dc8a562b1fb59953