Introduce a CLI tuleap-cfg to replace setup.sh
setup.el7.sh is better than setup.sh (c4/5/6) but it's still a shell script
that is a nightmare to debug and test. The goal of tuleap-cfg is to have
a well (unit) tested configurator made with modern tooling.

tuleap-cfg should be safe by default and idempotent so admins can run it
"with closed eyes".

Note: I put it in src/tuleap-cfg, I don't know if it's the right place.

Part of story #12639 a central tool to configure/setup/customize tuleap

Change-Id: Iee6003684bd53c616560996557842998ee612260