Add the wiki representation in the item representation of the docman
This is a part of the story #12444 Table : action on click

This feature will be used in the new docman to redirect the user to the wiki page if the item

How to test:
  - Make sure to have at least one wiki item in the docman.
  - Go to the API Explorer and test with route docman_items/:id or docman_items/:id/docman_items

You should see a new property which call "wiki_properties".
This property should be null if the item is not a wiki.

Change-Id: I7119682f47b8e246cbc0ee9b5ec350d4d6f496d0