request #12630: PR creation failure due to a Git reference creation failure is marked as a user error
If a Git operation fails when a user ask for the creation of the PR,
the error is marked as user error and the REST API returns a 400.
If a Git operation the error is not server side not client side therefore a 400
is not appropriate. It also means we get no log and we have no way to debug the
underlying issue.

Change-Id: Ia7d4e0a99c9731fb4360bc7496774c53da7d385c