Block item creation if there is an item being uploaded with the same name and parent
The goal is to put a lock a on item creation to let users uploading a document
the time to finish.

To test, you can add an entry to the plugin_docman_new_document_upload table.
The creation of an item with the same parent and name should be denied whatever
creation methods you use.

This is part of story #12612: upload document via REST route

Change-Id: I5df4b5233f73b09784bc65db75f20ad40e6d6854