request #12637: Document manager plugin should not crash when responding to an incorrectly formatted URL
The document manager plugin crashes when the group_id parameter is missing,
for example try to access to https://tuleap.example.com/plugins/docman/?action=search&id=29

Users are not expected to encounter this situation but even so, Tuleap should not
crash. It clutters logs on instances exposed publicly and it should and can be
gracefully managed.

Change-Id: Ic2f254473832c0d928208b23102358f358d6be5d