Display feedback if item is created
Part of story #12447: Modal new document for empty/wiki/link

Once the item is created, we should inform that the action is successful. This
is particularly needed when there is a lot of item in the current folder and we
don't know where the item has been created (typically under the page fold).

The module to add feedback should be used later also in the workflow, in
plugins/tracker/www/scripts/workflow-transitions/src/components/ChangeFieldConfirmationModal.vue
(worth a dedicated changeset).

Change-Id: Id87c539ae31cfc5dccb296d88c990f00283862cf