request #12818: Add metrics about Git LFS usage
This contribution adds a few metrics about the Git LFS usage:
 - number of accepted requests for the upload and download
of a Git LFS object
 - quantity of data received/transmitted to save/allow to download
a Git LFS object
 - number of locks created and deleted

For the data we sent, we count the data read from the object instead
of the data we really output. PHP does not seem to play well with a
stream filter and writing to php://output, the memory limit is reached.
Both values are anyway very close, in the worst case scenario one full
buffer (8192 bytes by default) can be lost.

Change-Id: I5fa5071b2500682279f7b5584276745bc022f334