Add a default LFS file preview in new Git UI
This is part of story #12325 use git-lfs objects in pull requests and web view

Only blob view is done for now. The color corresponds to the theme used
by the user.

The regexp provided in the detector class was built regarding the
current specification of the LFS file format[1].

The LFS text and SVG is directly handled by the Git plugin. We do this
choice because these two plugins are harldy linked and because it was
the best trade off regarding the feature / the code readability and the
security aspects.

[1] https://github.com/git-lfs/git-lfs/blob/v2.6.1/docs/spec.md#the-pointer

Change-Id: Id51c843e6de07431d4235e6c0e33ac6434350fca