Deal with LFS files in pullrequests
This is part of story #12325 use git-lfs objects in pull requests and web view

The LFS files are handled like binary files in PRs.

The file detection between source and destination file is the same as
the one defined for the binary files.
It will be enhanced in a dedicated commit.

Change-Id: I0b2f7a85594d1143d0e61969d3207ba8346d1c9d