Introduce the data structure to share document manager item ID between tables
The goal of this contribution is to have a way to reserve item ID
for documents being uploaded. This approach is obviously flawed and
in an ideal world we should not have to do that but in the current situation
it seems to be safest solution to not have ghost documents. Also, since
Tuleap still suports version of MySQL predating MySQL 5.6, we can not
use FULLTEXT indexes with the InnoDB engine. As a result we can not enforce
a constraint to be sure the item_id of the plugin_docman_item table has been
reserved.

To ease the migration process, since the table plugin_docman_new_document_upload
is not used at all in production environnements, the table is completly dropped
and recreated from scratch.

This is part of story #12612: upload document via REST route

Change-Id: I922f9f404c69b1b16e09d28ca722265b4a3aa991