Enable / disable Web UI Git LFS via Tuleap CLI
This is part of story #12326 forbid large git objects in favor of git-lfs

The option git_lfs_display_config is used to display or not the site
configuration.

It's available by default. The only way to disable it is to set the
option git_lfs_display_config to 0 with the Tuleap CLI.

Change-Id: I3d562b2ae1f36b961615e8f39089f1ed64d98a50