All write operations are blocked when using Gitolite2
Since 7734420567a1a5208e9349cf755bb9748aeea672 all write operations
done on a repo on a system using Gitolite2 are rejected with an error:

    can't find helper program for VREF/MAX_NEWBIN_SIZE/52428800
    error: hook declined to update refs/heads/master

Gitolite2 do not come with the VREF MAX_NEWBIN_SIZE so Gitolite is not able
to enforce it and rejects the update.

Since Tuleap Git LFS is only available with Gitolite 3 and that
is recommended for Gitolite 2 users to upgrade, this contribution only ensure
the updates are not blocked no limits are enforced.

This is part story #12326: forbid large git objects in favor of
git-lfs

Change-Id: Ic4689a0a638302e38faa257acaa51481299bbde2