Remove incorrect usage of $this
Two cases are encountered in the codebase:
  - $this is used in class static methods in the JPGraph library
which cannot work and raise a fatal error even on PHP 5.6. Code of
the impacted methods has been edited to throw a useful error instead
of fatal error, this will allow the JPGraph error handler to do its job.
  - PHPWiki configurator does multiple sketchy usages of $this which
cannot play well with PHP 7.1+ [0][1]. In the context of Tuleap, the
configuration cannot be changed as it is hardcoded by Tuleap itself. As
such, the whole configurator is dead code so this contribution removes
it instead of fixing it.

Part of request #12147: Fix PHP 7.2 warnings and errors

[0] https://wiki.php.net/rfc/this_var
[1] https://secure.php.net/manual/en/migration71.other-changes.php#migration71.other-changes.inconsistency-fixes-to-this

Change-Id: Ie738265442c4d4faa8f056bf6cb32eb2aec8a2b8