Remove dead code in the web UI update of the Git LFS max size
If the value can not be updated, an exception will be thrown so the feedback
won't added. Anyway there is little value to try to handle such case as it
either a developper mistake (e.g. incorrect SQL query) or a DB failure. There
is no way to properly recover from these issues so we should let the general
error handler do its job.

This is part of story #12326: forbid large git objects in favor of git-lfs

Change-Id: I941de6a2fc9c13909d5145dd81aaea3ad17a7ccc