Ensure we only work with a Git LFS max size as an integer value
Since ca15f6379565b2bf7d3d8a0d71035a64c149e839 the Git LFS max size
is stored in the forgeconfig table. The side effect is that the value
is now stored as a string instead of integer so we retrieve it from
the database as a string. Let's avoid the funny PHP implicit type casting
when comparing it to an integer by retrieving the value as integer.
This might save us from a painful debugging on a production instance.

This is part of story #12326: forbid large git objects in favor of git-lfs

Change-Id: I0950b875370f9c3cc1991c6571f0bf344fbc244d