Create a timetracking overview report
Part of request #11800 Know the time spent on different trackers on a single view

In this patch, database are upgraded to have timetracking_reports.
This report will be use to stock researches will be done on timetracking
Overview widget.

The already created widgets are upgraded to get their own report.
You can check on youre own database if it's work

Change-Id: If33ca775b28c7e7360b96e7e77302ce3cbcf0066