request #12721: Cryptography API is strict about types
The API was already enforcing types with specific checks,
since Tuleap does not support PHP 5.X anymore this can be managed
by the PHP engine directly.

Due to the sensitive nature of the operations handled, the internal
cryptography API is a prime candidate to use it. Implicit type conversion
does not mix well with cryptographic operations.

Change-Id: Iac4ae24c9b7bf9f3a8efd5879f695b5d0d93b766