Save modale should store links in database
This is a part of story #12447 Modal new document for empty/wiki/link

The link created and its version are now store in the database.

How to test :
  - Go to the new document UI
  - Click on 'New document'
  - Choose the link type
  - Fill the fields
  - click on 'Create document'

The modale should be closed and the new link is displayed.

NB: You can also test this feature with the API explorer
Change-Id: I1554dd5a039783c57c8e95c122dffacd54b37ab7