request #12719: Introduce a way to rebuild Gitolite configuration for all projects
This contribution adds a new command 'git:regenerate-gitolite-configuration'
to the Tuleap CLI.
This command can be useful when an administrator want to apply immediatly
a site wide change like the one introduced by story #12326. It can also be
useful if somehow a configuration is broken.

Configuration can be rebuilt for all projects with the option --all or for
a specific set of projects by specifying a list of project IDs.

Change-Id: I5f9f837fc956b1cf3b16eb4f76e9a7a4ddb78115