Make sure a file can be added to the document manager via the REST API
This contribution only extends what the REST test is doing. The
existing test was only reserving the space for a new file document but
the document was not really uploaded (and so not accessible to users), a
file is now really uploaded.

This only test the happy path for the upload itself. The tus "client" used
is very naive, no connectivity issue like the ones that can happen in real
life is expected. The responsability of testing this sort of scenarii should
not be held by a REST test about docman item creation but by the tus server
implementation itself.

This is part of story #10721: upload documents with drag'n drop

Change-Id: Ibfc0972b782b3b53a98f136a8b2a4ade1ebf20c2