Today computed value time must be set to 23:59:59
This is part of request #12558 Timezones issue with burndown and burnup graphs

The issue occured if today value is computed and there is cached value for
yesterday and both values are not the same. Only computed value must be
displayed for today.

Given that cached days time are 23:59:59, computed value for today must
also have this time set.

Change-Id: I1395dbf6b3bffe1d0171ce98d8fb71335b3ed240