Fix bad conflict resolution
Part of story #10721: upload documents with drag'n drop

In commit 9318344f43a9531bdf155a6993ea0c1b8d19acc0, it seems that a
conflict resolution leads to calling undefined mutation.

Try to add a document through the modal. There should be no error in the
console and the item should appear in the folder.

Change-Id: I22bd5aeea8332dddc8c21c661da0b396b0e1dc6a