Prepare the Tus server implementation to work with extensions
Extensions are some optional parts of the Tus protocol which may
or may not be needed depending on the context.

This technical change will allow us to add extensions in a similar
way to what is done to mark a file as available in the document manager.

No functional changes is expected.

Change-Id: Ic947c9969b60a8f0868d3c03c6e0f26237811189