The upload of document can be canceled
This contribution adds the support of the Tus termination extension [0]

To test, ask for the creation of new docman item file then call delete
on the upload route. The lock on the docman item should be removed,
another docman item with the same name can be created.

This is part of story #10721: upload documents with drag'n drop

[0] https://tus.io/protocols/resumable-upload.html#termination

Change-Id: I3971944f7e82d81773e9d64e181423ba6989ee15