Add application/gzip to the list of icons
Part of story #10721: upload documents with drag'n drop

Create a .tar.gz file, drop it in the docman. The icon that is displayed
during the upload should be the same than when the upload is finished.

Browser uses application/gzip, but server detects the file as
application/x-gzip. application/gzip is a recent mime type that has been
added in 2012, see RFC 6713:
https://tools.ietf.org/rfc/rfc6713.txt

Change-Id: Ia299f318595a3183544c4acacf5dc298bc1a24c4