Prevent concurrent uploads of the same document by the same user
Otherwise this means that a user can try to upload a same file in
different contexts (e.g. two browsers, different browser tabs...)
and weird errors can happen.

The conflict management is already partially handled by the fact that
a client has to provide an upload offset equal to the current offset
of the resource.

The tus protocol does not have explicitly something to deal with the lock
of a resource, however official implementations handles it with a HTTP 423
code [0][1], therefore the same solution has been chosen.

This is part of story #10721: upload documents with drag'n drop

[0] https://github.com/tus/tusd/blob/0.11.0/unrouted_handler.go#L59
[1] https://github.com/tus/tus-js-client/blob/v1.5.2/lib/upload.js#L350

Change-Id: I3767f1902fac3c2cd0d18759068abe4aea2a635b