request #12760: Ensure the deployed Git post receive hook is up to date on "old" instances
On old instances, we might have deployed a file that is not a symlink and
therefore might not be up to date. If the file is not up to date, it might
not be in sync with the sudoers field and not be executed anymore. This
contribution ensure it is always up ot date after an upgrade.

Change-Id: I84086af5772e24d020f3721c6e2d6b6a7a98ec27