Workflow actions must not be processed when workflow is not active
This fixes request #12775 Workflow post actions are executed even on disabled workflow
This is also part of story #12189 prevent field modification according to state

A new flag has been added to the workflow, so that current behaviour for
exiting workflows does not change.

All new and newly activated workflows will take into account activation
state to process or not pre-condition and post-actions.

The actions concern both post actions and checks done before the
artifact update.

How to test:
- ForgeUpgrade should flag all unused workflows with post actions to
legacy
- When your workflow is disabled and no more legacy, with this patch post actions should no
longer be executed.
- When your workflow is disabled and legacy, post actions should be executed.

A message is displayed in the legacy workflow view. A dedicated patch
will be done for the new Workflow UI.

This will also be documented in the Tuleap documentation.

Change-Id: Idb4176c126c44f80fd09b8e43560b35282411f49