The tus server should mark a file as finished only when all data has been copied
Currently if the client only sends a chunk of the whole document, the partially
uploaded document will be considered has finished. In the case of the document
manager this means an incomplete file can be made downloadable by other users.

This is part of story #10721: upload documents with drag'n drop

Change-Id: Ie55186c742254c84af1ff64c849ad9701cb230a1