Implement lock verify
This change is part of story #12323

This feature is secured by a feature flag in Forge Config:
feature_gitlfs_lock_api

It can be tested by calling directly the API.

/!\ Following discussions, although pagination is described in the
Git LFS spec, it has been removed from this commit due to the fact
that the same authentication token is used for going through it which
is incompatible with our one-shot tokens implementation. It will be
implemented later on if ever needed. /!\

Change-Id: Ief3825d198863848d7bbf40b970bc9c4971e271b