Mitigate incoherent states if an issue happend while creating a new file document
Some of the tables of the docman plugin use the MyISAM engine. Unfortunately,
the MyISAM engine does not support transactions which means we might have
data in a incoherent state if an issue happen while creating a new file document.

This contribution tries to mitigate this fact, no functionnal changes are expected
for the nominal case.

This is part of story #10721: upload documents with drag'n drop

Change-Id: I5b8db128cab8ec1ef6ce397144dcabea8b99cacf