Do not display only once the error modal on drop
Part of story #10721: upload documents with drag'n drop

Try to drop a too big file or a file that already exists: an error modal
is displayed. Close the modal and retry immediately with the same file:
the error modal is displayed again.

Bug introduced by commit 0fbda56fa47ef4303b9d190648cacb0315353eb0.

In VueJS, $emit does not bubble the event, you have to manually do it
(or use an external lib but it is not worth it).

Change-Id: Ib9ca64237cb2fd8f3168a8a3b109a5817eb9ed06