Workflow is_legacy information is returned by REST route
This is part of request #12775 Workflow post actions are executed even on disabled workflow
This is also part of story #12189 prevent field modification according to state

The information is now returned by the REST route.
This information will be used by the Vue.js app to display the warning
message if needed.

Change-Id: Icdde4290f791fb160fc9121a0df5d9b0d5d0c588