Ignore files in errors for global progress bar
Part of story #10721: upload documents with drag'n drop

Drop a file that takes ages to upload.
Now in FileUploadController, throw new \Exception() in process to
simulate an upload error for the next dropped file.
Drop a small file (few bytes so that you don't wait too long).

The upload modal appears, indicating that there is a big file still
uploading (maybe still at 0%) and the small file in [Upload error].
Close the modal.

The global progress bar should display 0% and not 50%.

Change-Id: I8aceb530c50ed399e8eb9b91e1b0cc32f23ec575