request #12828: Default wiki page names must be escaped before being used in a SQL query
Otherwise the DB is not able to encode the strings correctly and warnings
or errors might happen when executing the queries.

Even without encoding issues related to the character set being used,
any inputs should not be blindly concatened into a SQL query.

In a case where defaults are applied to the DB (a standard dev instance for example)
this contribution does not change anything functionally. In other cases it prevents
the impacted SQL queries failures.

Change-Id: Ie6b16efb48847ecca9985c2a9f1a51b0d9251dcb