Workflow legacy transitions can be deactivated with REST
This is part of request #12775 Workflow post actions are executed even on disabled workflow
This is also part of story #12189 prevent field modification according to state

A new query can be provided to PATCH trackers/:id like:
{ "workflow": { "is_legacy": false } }

It's not possible to switch a worklow to a legacy state.
If the user set is_legacy to true, an error will be thrown.

Change-Id: I2eaa9d99d4178e038d0b9cd79d9c99c96884cdfc