request #12872: Document manager should not trust the file size given by the end user
The file size given by the end user should not be trusted blindly. The given value
is used in the Content-Length header when downloading the file. As a result,
if a wrong size is given the download will considered as corrupted or truncated
by browsers/HTTP clients.

This contribution fixes the root cause (the given size is not trusted anymore)
and ensure files that has been created with a wrong size can still be downloaded.

Change-Id: Ic109ac2fd203b147ec41db49d134c6acd2e3c42c