Bump REST tests to PHPUnit 8 and prepare unit tests to the move
Unit tests can not be moved to PHPUnit 8 due to the incompatibility of
mockery/mockery.

The main changes are the addition of a return type
on some PHPUnit methods and the deprecations of the
annotations @expectedException...

The update has highlighted a few issues in the REST tests.
For a lot of tests the status code of the response was not verified
due to the thrown exception, we had a lot of dead code. In order to solve
that, the HTTP client does not throw exception on errors anymore. This change
will also help to a future upgrade to a PSR-18 HTTP client.

Two REST tests has been found has broken, they must be fixed in dedicated contributions:
 * TrackersTests::testGetDeletedTrackerReturnsError
 * DocmanItemsTest::testPostReturns403WhenPermissionDenied

Our custom assertions are not compatible with PHPUnit 8, they has been removed since
the same thing could be achieved without the custom assertions and without more code
duplication.

This is part of request #12959: Bump to PHPUnit 8

Change-Id: Ibc3605978cbc94aeddde28b61795b6efccdee04f