request #12961: Some PHPUnit unit tests are never played
The commit ccff866a14754e24d01982da6b8d69164385a31f has been reverted
for now. The Docker images running the PHPUnit unit tests does not come
with Perl (and they probably should not). Also LanguageFilesTest is not
really a unit test. This check should probably be isolated in a dedicated
step in the test pipeline in a similar fashion of what is done for the
po files.

Change-Id: I2cd91ff97472096ec736969ddb412f82b99dd7f9