request #12963: Project PHPWiki pages might not be displayed if some metadata are missing
When the wiki_group_list table contains no information about a project,
the list of pages of this wiki is not displayed.
This is caused by the fact the project ID is automagically set globally
by the instanciation of some wiki pages when processing the wiki_group_list
table information. This contribution sets explicitely the project ID before
displaying the list of the wiki pages.

The wiki_group_list table is not supposed to be empty for a project but it
seems it might have happened in a few projects on productions instances. Besides
the display issue is not a major issue, PHPWiki fallbacks to the default information
when nothing is found.

Change-Id: I2962ab968e136e3ff698a50b353bd290a10e6f18