Allow usage of UTF-8 in sentences to be translated
Part of story #12553: see flags that categorize my project in all project pages

In the previous commit, 'Choose a category' should be followed by '…' (U+2026)
instead of three dots '...'. However our target `make generate-po`
complained about the lack of UTF-8 charset header in the generated .pot file.

In order to not clutter the review of the previous one, this is fixed in
a dedicated commit. We take the opportunity to remove the .pot files
from sources because they are generated temp files.

Change-Id: I14cbd871239dbbc2dc3255dd7b98c219301f9d3b