Admin can choose nb max values for a category
Part of story #12553: see flags that categorize my project in all project pages

As site admin, edit a top level category and adjust the "Nb maximum of
allowed values". The chosen value is reflected in project » admin »
categories UI.

Bonus:
- align inputs in the modal and move description field as requested on
the mockups.
- fix request #11583

Those two bonuses can be extracted into dedicated commits if you think
it is better for the review.

Change-Id: I6d92dfa7280cf498d29044072d6c54126fd6158e