request #12990: Displaying a cumulative flow chart with a "deprecated browser" crashes the page
Reproduction steps:
 1. Add a cumulative flow chart of a tracker to a dashboard
 2. Try to access to the dashboard with IE11 in compatibility mode

Issue is present since 79247604ac8a590c04489ebd9e0b8c3263fa2612.

The detected browser is not considered anymore to know if we can use D3 or not.
If the browser is not able to display some D3 graphs other parts of Tuleap are likely
to break anyway.

Change-Id: I857b6637150109f974bbba7a70557d1717e3e953