request #12991 Project admin can't see all workflow transitions
As project admin, using the new Workflow Admin UI for convenience,
it can be done by REST alone,

Given a user group "MyUserGroup" that I am not member of,
When I edit an existing transition and select "MyUserGroup" for the user groups
that have permission to process the transition (in pre-conditions),
Then upon refresh I will no longer see that transition.

The REST API GET /trackers/:id will omit it, so the new UI will not list it.

Instead of this, as I am project admin and am in charge of configuring this,
I should bypass permissions and see all transitions.

In addition, all tracker administrators could be able to process any transition.

Change-Id: Ia5babc76ded0a7b0907e639bc0367308a8f4448c