request #12992 Tracker workflow pre conditions not checked at artifact creation.
If you defined ugroups that are allowed to process a transition from (New artifact)
and/or if you defined some fields that must not be empty in this transition,
at artifact creation, these pre conditions are currently not checked.

To fix this, we will add a forgeupgrade to map the current behaviour
(access to all_users and no field empty) and the validation at
initial changeset creation.

Change-Id: Ic3741ea6bed442f967877ab378069d14d674bc1f