request #12993: Update modal of the directory ugroup binding looks for the group using the CN attribute of the users
Use the common name of the group instead of the one of the users.

Otherwise we might end up searching on a non existing LDAP group,
if the attribute CN of the users does not exist on the group objects
the search for the group fails. We should get the same CN than the one
we use when we search the group.

Change-Id: I19f030732f74d76d600307b1577367f5e7c1b1a5